+++ spec/unit/resources/interconnect/set_power_state_spec.rb
@@ -9,6 +9,11 @@ describe 'oneview_test::interconnect_set_power_state' do
     expect_any_instance_of(OneviewSDK::Interconnect).to receive(:patch).with('replace', '/powerState', anything)
     expect(real_chef_run).to set_oneview_interconnect_power_state('Interconnect2')
   end
+
+  it 'fails if the resource is not found' do
+    expect_any_instance_of(OneviewSDK::Interconnect).to receive(:retrieve!).and_return(false)
+    expect { real_chef_run }.to raise_error(RuntimeError, /not found/)
+  end
 end
 
 describe 'oneview_test::interconnect_set_power_state_invalid' do