+++ spec/unit/resources/logical_interconnect/activate_firmware_spec.rb
@@ -17,4 +17,9 @@ describe 'oneview_test::logical_interconnect_activate_firmware' do
       .with('Activate', instance_of(OneviewSDK::FirmwareDriver), anything).and_return(true)
     expect(real_chef_run).to activate_oneview_logical_interconnect_firmware('LogicalInterconnect-activate_firmware')
   end
+
+  it 'fails if the resource is not found' do
+    expect_any_instance_of(OneviewSDK::LogicalInterconnect).to receive(:retrieve!).and_return(false)
+    expect { real_chef_run }.to raise_error(RuntimeError, /not found/)
+  end
 end