+++ spec/unit/resources/logical_interconnect/update_settings_spec.rb
@@ -15,4 +15,9 @@ describe 'oneview_test::logical_interconnect_update_settings' do
     expect_any_instance_of(OneviewSDK::LogicalInterconnect).to receive(:update_settings).and_return(true)
     expect(real_chef_run).to update_oneview_logical_interconnect_settings('LogicalInterconnect-update_settings')
   end
+
+  it 'fails if the resource is not found' do
+    expect_any_instance_of(OneviewSDK::LogicalInterconnect).to receive(:retrieve!).and_return(false)
+    expect { real_chef_run }.to raise_error(RuntimeError, /not found/)
+  end
 end